requirements.txt
Django==2.2.12
djangorestframework==3.13.1
